refactor: location package로 이름 변경
